Refactor author module

+++ src/main/java/pl/edu/pw/pamiwapi/services/AuthorService.java
@@ -77,8 +77,20 @@ public class AuthorService {
                 .build();
     }
 
-    public void delete(int id) {
+    public ServiceResponse<Author> delete(int id) {
+        if (!repository.existsById(id)) {
+            return ServiceResponse.<Author>builder()
+                    .wasSuccessful(true)
+                    .message("No author to delete.")
+                    .build();
+        }
+
         repository.deleteById(id);
+
+        return ServiceResponse.<Author>builder()
+                .wasSuccessful(true)
+                .message("The author was successfully deleted.")
+                .build();
     }
 
     public Author mapFromDto(AuthorDto dto) {
